restructuring the map-trait a bit
